Directories refactored
