--- parser/credentials.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "dataoxparser",
-  "private_key_id": "55092a04472036433f0474e59ddfed1b9fc879d6",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQDZeGbmaAIVuGoG\nBTte0fRXyEEB5ZcpJDmfwA5BprGyaPHykDfNyxrdmeF1VP7TpJfxe+6qSEQLl81o\n/E9CHt/Lo3P7HKyjdwx7sfcsSjw0hLpN2XktvumG9U3j4shejMISpjB9vcGFvsE5\n4EttAc0CQq9AJWXVmuu7aU/flOd8hjd6QK2EAR24ARGycshEaRKsMZ4zW3EW+0on\nx8KZ0PKwVu5pdUrotiTeS7Pvsdi16AjR/2qKA4L1KujkQA0AItblr/xy+xUA98Jx\nZb0HYBp5+QcHk6glOyRfFtucKbuYfN5bfO/XviQddmnKRaFvEasXKJfi0pIA9HzC\nSrdYdlg9AgMBAAECggEABEDy8x1kTD93F3B0jyMlMXuRxfM5mzNOfnDxXNhhXNK2\nFNJvicxBiMWEEvL09XfTnwR93RsC9bhKhMiwPJdbaK5lGN7TnT9S+uJp9skdV1uk\nENzCayImWDgl7C5MUNXvZkT8HG5UDp7C3ppw1aXo3o5Zj82+qa7ZK8L6FHuWcY/3\nR+XYAVG+dzqpt4l99SPpCF6f5b5LPNXNLeogniVI3CVX7Agj1nf5EnbCgm31Ag9U\nPE1LDWUYY59oU87O5em/+xeiRZ676M9oCjP71h09ni9HX1QoR49bAhIbLEH9ek15\n9wwErpY9JRGhmmOxqEweTzqTttMtpQEy5k4XADte4QKBgQD0msZ2Nlzej5PtGTqA\nWORpQZJsEAUPxTJu4dzZG9o00azwwB0Y4s1ek14yOlfevAuss+Kj/l84bxaHqh8U\n3W+PFL6V6YCHtQ+1recAldn0Yxccr3Zz5n0hR8F1M3M3pk7zNuyb5gPySkDcMTaq\ntxxYqnLGkfdMqpfk9RXMVQ3FsQKBgQDjmgP9rRwr86ivc9BGwVjkZJluVcjGLxaZ\nJede8W6VKgLx+9mCcSXgL9dx2qSO6VucorrUgdOp+nzLd17qiA9syvgJrx8kSJaU\nrhUGbxy4BxYqm3fZDpYyk06FgpT6ZEJaytw6vZm8iZ+KmSOQVRHl762zetDVG3r8\nq8xkWU6CTQKBgQDUqOO2zti7HrULRA3anG4pHLtCfEMKTPXl9TRfBZoWpdzniv9B\nLZ5DhXg1zweqPG8h/VskHPmIRPyfgkZ3i3V8PhrclyAJ3F8GI2Vc2uudT7E+XUZs\nyY0srAgD73FShqLJTY00lSwFMI1nYtoWXhCk/bu/7BPHceGRS4tiFOheYQKBgQDR\na8+sRPKdBm/6MKg90n4GeQIsYcaBhKZYeGXyGFhV6/gVzFSRD6d4Js0oN7Qew13G\nPH2IN4UY7CCT3tx5FISM6B1eW8Jd0mSrnZo+91IHio282pxmHmPZWW1wr3yBAGuF\nuw71BaxFrKQeFVaB4aS8W9MKSEGLNOPvReylkP8KKQKBgQDogHlYMQIzznvX0PU8\n9dUy2r0Wd4FfoYV6wdr0W/velETNgn/mfMlmrAZKV/QZ+M/agEr2/Jgx3MMY6NXl\ngSd/nbc+vGmSTNK2XvLVzm2tcNu3eMdftJj30nayWTH7YagcXnu7zmNagdgQdmwn\n+nQ+I10YCtyZuDX8REf26p7pwA==\n-----END PRIVATE KEY-----\n",
-  "client_email": "dataoxparseraccount@dataoxparser.iam.gserviceaccount.com",
-  "client_id": "114632740101704211219",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/dataoxparseraccount%40dataoxparser.iam.gserviceaccount.com"
-}
